Reworked app flow -> log in no longer required for most actions. Log in screen is no longer main activity
