+++ app/src/main/java/com/example/mobproj2020new/DatabaseHandler.java
@@ -10,6 +10,7 @@ import androidx.annotation.NonNull;
 
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentReference;
@@ -36,19 +37,17 @@ public class DatabaseHandler {
     private String mPhone = "empty";
     public static final int REQKEY = 1212;
     private StorageReference storageRef;
-    private FirebaseStorage fbs;
 
 
-    public void init(FirebaseUser user)
-    {
-        uid = user.getUid();
-        fbs = FirebaseStorage.getInstance();
-        storageRef = fbs.getReference();
-        mUsersDocRef = FirebaseFirestore.getInstance().collection("users").document(uid);
-    }
 
     public void setUserCreationInfo(String fname, String lname, String phone)
     {
+        String uid = FirebaseAuth.getInstance().getCurrentUser().getUid();
+        if(uid.equals("")) {
+            return;
+        }
+        mUsersDocRef = FirebaseFirestore.getInstance().collection("users").document(uid);
+
         User user = new User(fname,lname,phone);
         mUsersDocRef.set(user); // add on success, on failure event listeners for checking for errors if needed
     }
@@ -73,25 +72,15 @@ public class DatabaseHandler {
         });
     }
 
-    public void getPhone()
-    {
-        mUsersDocRef.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
-            @Override
-            public void onComplete(@NonNull Task<DocumentSnapshot> task) {
-                if(task.isSuccessful()) {
-                    DocumentSnapshot doc = task.getResult();
-                    if(doc.exists()) {
-                        // Creating User object based on document and getting phone string
-                        String phone = doc.toObject(User.class).getPhone();   // Currently only printing this value
-                        Log.d("HALOOOO", ("phone number from db: " + phone));
-                    }
-                }
-            }
-        });
-    }
+
 
     public void checkProfileCreated(Context context)
     {
+        String uid = FirebaseAuth.getInstance().getCurrentUser().getUid();
+        if(uid.equals("")) {
+            return;
+        }
+        mUsersDocRef = FirebaseFirestore.getInstance().collection("users").document(uid);
         final Context varContext = context;
         mUsersDocRef.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
             @Override
@@ -103,12 +92,15 @@ public class DatabaseHandler {
                         boolean created = doc.toObject(User.class).getProfileCreated();
                         if(created)
                         {
-                            Intent intent = new Intent(varContext, ChoosePickUpOrTransportationActivity.class);
+                            FirebaseHelper.loggedIn = true;
+                            /*
+                            Intent intent = new Intent(varContext, MainActivity.class);
                             intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                            varContext.startActivity(intent);
+                            varContext.startActivity(intent);*/
                         }
                         else
                         {
+                            FirebaseHelper.loggedIn = false;
                             Intent intent = new Intent(varContext, EditProfileActivity.class);
                             intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
                             varContext.startActivity(intent);
@@ -121,6 +113,11 @@ public class DatabaseHandler {
 
     public void setProfileCreated(final boolean value)
     {
+        String uid = FirebaseAuth.getInstance().getCurrentUser().getUid();
+        if(uid.equals("")) {
+            return;
+        }
+        mUsersDocRef = FirebaseFirestore.getInstance().collection("users").document(uid);
         mUsersDocRef.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
             @Override
             public void onComplete(@NonNull Task<DocumentSnapshot> task) {
@@ -136,6 +133,9 @@ public class DatabaseHandler {
 
     public void putImageToStorage(Uri imageUri)
     {
+        FirebaseStorage fbs;
+        fbs = FirebaseStorage.getInstance();
+        storageRef = fbs.getReference();
         StorageReference ppRef = storageRef.child("profpics/"+uid);
         UploadTask upTask = ppRef.putFile(imageUri);
         // Add listeners for fail/complete/success?
