+++ app/src/main/java/com/example/mobproj2020new/MainActivity.java
@@ -2,125 +2,112 @@ package com.example.mobproj2020new;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.constraintlayout.widget.ConstraintLayout;
 
+import android.app.AlertDialog;
+import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
-
-import android.graphics.Color;
-
-import android.provider.ContactsContract;
-import android.util.Log;
-import android.widget.EditText;
-
 import android.os.Bundle;
+import android.util.Log;
+import android.view.Gravity;
 import android.view.View;
-import android.widget.Button;
-import android.widget.Toast;
 
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 
-
-public class MainActivity extends AppCompatActivity {
-
-    private FirebaseAuth mAuth;
-    private DatabaseHandler db;
-
-    private final String TAG = "HALOOOOO";
-
-    EditText userEdit;
-    EditText passEdit;
-
-    String userNameStr = "";
-    String passwordStr = "";
+public class MainActivity extends AppCompatActivity{
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        mAuth = FirebaseAuth.getInstance();
-        db = new DatabaseHandler();
-
-        userEdit = findViewById(R.id.usernameEdit);
-        passEdit = findViewById(R.id.passwordEdit);
+        setContentView(R.layout.activity_choose_pick_up_or_transportation);
 
         FirebaseAuth.AuthStateListener als = new FirebaseAuth.AuthStateListener() {
             @Override
             public void onAuthStateChanged(@NonNull FirebaseAuth firebaseAuth) {
-                if(mAuth.getCurrentUser() != null)
+                if(FirebaseAuth.getInstance().getCurrentUser() != null)
                 {
                     CheckProfileCreated();
                 }
                 else
                 {
-                    // User not logged in anymore...
-                    // Maybe return to MainActivity here?
+                    FirebaseHelper.loggedIn = false;
                 }
             }
         };
-        mAuth.addAuthStateListener(als);
+        FirebaseAuth.getInstance().addAuthStateListener(als);
     }
 
     private void CheckProfileCreated()
     {
-        Log.d("HALOOOOOOOOOOOOOOOOO", "Taalla ollaan");
-        db.init(mAuth.getCurrentUser());
+        DatabaseHandler db= new DatabaseHandler();
         db.checkProfileCreated(getApplicationContext());
     }
 
-    //Login button press
-    public void login(View V){
-        String email = userEdit.getText().toString();
-        String password = passEdit.getText().toString();
-
-        if(email.equals(""))
-        {
-            noEntry(userEdit);
-        }
-        else if(password.equals(""))
-        {
-            noEntry(passEdit);
-        }
-        else {
-            mAuth.signInWithEmailAndPassword(email, password)
-                    .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
-                        @Override
-                        public void onComplete(@NonNull Task<AuthResult> task) {
-                            if (task.isSuccessful()) {
-                                // Sign in success, update UI with the signed-in user's information
-                                Log.d(TAG, "signInWithEmail:success");
-                            } else {
-                                // If sign in fails, display a message to the user.
-                                Log.w(TAG, "signInWithEmail:failure", task.getException());
-                                Toast.makeText(MainActivity.this, "Authentication failed.",
-                                        Toast.LENGTH_SHORT).show();
-                            }
+    //-----------Applications settings button------------//
+    public void AppSettings(View v) {
+        final String[] itemList = {"Settings", "My Profile", "About", "Sign Out"};
+        AlertDialog.Builder builder = new AlertDialog.Builder(this);
+        builder.setItems(itemList, new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialog, int which) {
+                switch (which){
+                    case 0:
+                        //Log.d("SWAG", "onClick: settings");
+                        break;
+                    case 1:
+                        //Log.d("SWAG", "onClick: My Profile");
+                        if(FirebaseHelper.loggedIn) {
+                            DatabaseHandler db = new DatabaseHandler();
+                            db.GoToProfile(MainActivity.this, FirebaseAuth.getInstance().getCurrentUser().getUid());
+                        } else {
+                            FirebaseHelper.GoToLogin(getApplicationContext());
                         }
-                    });
-        }
+                        break;
+                    case 2:
+                        //Log.d("SWAG", "onClick: About");
+                        break;
+                    case 3:
+                        //Log.d("SWAG", "onClick: Sign Out");
+                        if(FirebaseHelper.loggedIn) {
+                            FirebaseAuth.getInstance().signOut();
+                            Intent intent = new Intent(getApplicationContext(), LogInActivity.class);
+                            intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+                            startActivity(intent);
+                            break;
+                        }
+                }
+            }
+        });
+        AlertDialog settingsDialog = builder.create();
+        settingsDialog.show();
+        settingsDialog.getWindow().setDimAmount(0);
+        settingsDialog.getWindow().setLayout(420, ConstraintLayout.LayoutParams.WRAP_CONTENT);
+        settingsDialog.getWindow().setGravity(Gravity.TOP | Gravity.RIGHT);
+        settingsDialog.getWindow().setBackgroundDrawable(getDrawable(R.drawable.settings_dialog_background));
     }
 
-    //SignUp button press
-    public void signUp(View v){
-        Intent signUpIntent = new Intent(this, SignUp.class);
-        startActivity(signUpIntent);
+    //----------------Button BookedTrips----------------//
+    public void SelectBookedTrips(View v){
+        DatabaseHandler db = new DatabaseHandler();
+        db.GoToProfile(MainActivity.this, FirebaseAuth.getInstance().getCurrentUser().getUid());
     }
 
-    //execute if username or password is empty
-    public void noEntry(EditText et){
-        et.setText("");
-        et.setHintTextColor(Color.parseColor("#B75252"));
-        if (et.getId() == R.id.usernameEdit) userEdit.setHint("Username*");
-        else passEdit.setHint("Password*");
+    //---------------Button OfferTrips---------------//
+    public void SelectOfferedTrips(View v){
+        DatabaseHandler db = new DatabaseHandler();
+        db.getMatchingRoutes(30f, 64.99960786f, 25.50759315f, 62.27942323f, 25.7258606f);
     }
 
-    //Exit app with pressing back putton on your phone
-    @Override
-    public void onBackPressed() {
-        super.onBackPressed();
-        return;
+    //-------------Button Get A Ride----------------//
+    public void SelectGetARide(View v){
+        Intent GetARideIntent = new Intent(MainActivity.this, GetRideActivity.class);
+        startActivity(GetARideIntent);
     }
 
+    //------------Button Offer A Ride--------------//
+    public void SelectOfferARide(View v){
+        Intent intent = new Intent(MainActivity.this, RouteActivity.class);
+        startActivity(intent);
+    }
 }
