+++ app/src/main/res/layout/activity_choose_pick_up_or_transportation.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background_login"
-    tools:context=".ChoosePickUpOrTransportationActivity">
+    tools:context=".MainActivity">
 
     <ImageView
         android:id="@+id/btnSettings"
